Revert "Sidebar card filters"
