--- src/css/AdvancedOptions.css
@@ -1,69 +1,0 @@
-.advancedOptionsButton {
-    font-size: 2em;
-    margin-left:1em;
-    border: 1px solid var(--clr-txt);
-    border-radius: 5px
-}
-
-.advancedOptionsButton:hover {
-    cursor: pointer;
-    filter: brightness(0.8);
-}
-
-.advancedOptionsModal {
-    text-align: left;
-    background-color: rgb(70, 70, 70);
-    border-radius: 10px;
-    line-height: 1.4285em;
-    padding: 20px;
-}
-
-.closeModalButton {
-    width: 100%;
-    height: 2.5em;
-}
-
-.closeModalButton button {
-    float: right;
-    border-radius: 3px;
-    width: 2.5rem;
-    height: 2.5rem;
-    font-size: 2em;
-    border: 1px solid var(--clr-txt);
-}
-
-.closeModalButton button:hover {
-    cursor: pointer;
-    filter: brightness(0.8);
-}
-
-.closeModalButton i {
-    margin: 0;
-    width: auto;
-    height: auto;
-}
-
-.advancedOptionsSection {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 15px;
-    margin-bottom: 15px;
-    align-items: flex-start;
-}
-
-.advancedOptionsSection .rarityLabel {
-    margin-bottom: 15px;
-}
-
-.advancedOptionsSection .colorBox{
-    padding: 0;
-}
-
-.advancedOptionsSection .colorList {
-    padding-top: 0;;
-}
-
-.advancedOptionsSection .reset-button {
-    height: 35px;
-    align-self: flex-end;
-}
