+++ src/css/CustomDropdown.css
@@ -7,7 +7,7 @@
 }
 
 .dropdown-selected{
-    padding: 2.5px;
+    padding: 2%;
     border: 1px solid var(--clr-txt);
 }
