--- src/css/DBCardFilters.css
@@ -1,12 +1,0 @@
-.DBCardFilters {
-    font-size: 0.8rem; 
-}
-
-.DBCardFilters .reset-button {
-    font-size: 1em;
-}
-
-.DBCardFilters .advancedAndReset {
-    display: flex;
-    
-}
