+++ src/css/DBSidebar.css
@@ -1,6 +1,5 @@
 #DBSidebar {
     width: 20%;
-    min-width: 225px;
+    min-width: 200px;
     border-right: 2px solid var(--clr-txt);
-    padding: 5px;
 }
